+++ .gitignore
@@ -162,4 +162,4 @@ cython_debug/
 examples/
 pytest.ini
 .DS_store
-test_sessions/
+*.db
