+++ requirements.txt
@@ -2,3 +2,4 @@ aiohttp==3.9.5
 line-bot-sdk==3.11.0
 fastapi==0.111.0
 uvicorn==0.30.1
+SQLAlchemy==2.0.31
